+++ tests/testthat/test-expr_t_twosample_nonparametric.R
@@ -8,7 +8,8 @@ test_that(
     # ggstatsplot output
     set.seed(123)
     using_function <-
-      statsExpressions::expr_t_nonparametric(
+      expr_t_twosample(
+        type = "np",
         data = mtcars,
         x = am,
         y = wt,
@@ -82,7 +83,8 @@ test_that(
       ))
 
     # expect error
-    expect_error(suppressWarnings(statsExpressions::expr_t_nonparametric(
+    expect_error(suppressWarnings(expr_t_twosample(
+      type = "np",
       data = iris,
       x = Sepal.Length,
       y = Species
@@ -91,7 +93,8 @@ test_that(
     # ggstatsplot output
     set.seed(123)
     using_function2 <-
-      suppressWarnings(statsExpressions::expr_t_nonparametric(
+      suppressWarnings(expr_t_twosample(
+        type = "np",
         data = df_bird,
         x = type,
         y = length,
@@ -141,7 +144,8 @@ test_that(
   desc = "dataframe",
   code = {
     expect_s3_class(
-      statsExpressions::expr_t_nonparametric(
+      expr_t_twosample(
+        type = "np",
         data = dplyr::filter(movies_long, genre == "Action" | genre == "Drama"),
         x = "genre",
         y = rating,
@@ -193,7 +197,8 @@ test_that(
     # incorrect
     set.seed(123)
     expr1 <-
-      statsExpressions::expr_t_nonparametric(
+      expr_t_twosample(
+        type = "np",
         data = df,
         x = condition,
         y = score,
@@ -204,7 +209,8 @@ test_that(
     # correct
     set.seed(123)
     expr2 <-
-      statsExpressions::expr_t_nonparametric(
+      expr_t_twosample(
+        type = "np",
         data = dplyr::arrange(df, id),
         x = condition,
         y = score,